requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyspark>=3.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.8.0
apache-airflow>=2.7.0
jupyter>=1.0.0
